Final Commits before project turn-in

+++ routes/user/index.js
@@ -17,9 +17,7 @@ router.post("/createUser", async (req, res) => {
       createdAt: new Date(),
       updatedAt: new Date(),
     });
-    res.render("./user/user.ejs", {
-      user: test,
-    });
+    res.redirect("/login");
   });
 });
 
@@ -27,8 +25,12 @@ router.get("/createUser", async (req, res) => {
   res.render("./user/createUser.ejs");
 });
 
-router.post("/:id", async (req, res) => {
-  const userID = parseInt(req.params.id);
+router.post("/deleteUser", async (req, res) => {
+  if (!req.user) {
+    res.redirect("/login");
+    return;
+  }
+  const userID = req.user.id;
 
   try {
     const deletedUser = await user.destroy({
@@ -38,17 +40,64 @@ router.post("/:id", async (req, res) => {
     if (deletedUser === 0) {
       return res.status(404).json({ error: "User not found" });
     }
-
-    res.json({ message: "User deleted successfully" });
+    req.logout(function (err) {
+      if (err) {
+        return next(err);
+      }
+      res.redirect("/login");
+    });
   } catch (error) {
+    console.error(error);
     res
       .status(500)
       .json({ error: "An error occurred while deleting the user" });
   }
 });
 
-router.get("/:id", async (req, res) => {
-  res.render("./user/deleteUser.ejs");
+router.get("/updateUser/", async (req, res) => {
+  if (!req.user) {
+    res.redirect("/login");
+    return;
+  }
+  const userID = req.user.id;
+
+  try {
+    const userToUpdate = await user.findByPk(userID);
+
+    if (!userToUpdate) {
+      return res.status(404).json({ error: "User not found" });
+    }
+
+    res.render("./user/user.ejs", {
+      user: userToUpdate,
+    });
+  } catch (error) {
+    res
+      .status(500)
+      .json({ error: "An error occurred while fetching the user" });
+  }
+});
+
+router.post("/updateUser", async (req, res) => {
+  if (!req.user) {
+    res.redirect("/login");
+    return;
+  }
+  const userID = await req.user.id;
+  const { name, email, password } = req.body;
+  try {
+    const userToUpdate = await user.findByPk(userID);
+    bcrypt.hash(password, saltRounds, async function (err, hash) {
+      userToUpdate.set({ name: name, email: email, password: hash });
+      await userToUpdate.save();
+    });
+  } catch (error) {
+    console.error(error);
+    res
+      .status(500)
+      .json({ error: "An error occurred while updating the user" });
+  }
+  res.redirect("/user/updateUser");
 });
 
 module.exports = router;
